src/pdf-lbr-output-bib.pl: implement BibTeX field filter

+++ src/pdf-lbr-output-bib.pl
@@ -43,7 +43,7 @@ B<pdf-lbr-output-bib> - Output BibTeX bibliographic metadata from PDF files.
 
 B<pdf-lbr-output-bib> B<--help>|B<-h>
 
-B<pdf-lbr-output-bib> [ B<--clipboard>|B<-c> ] [ B<--max-authors>|B<-m> I<count> [ B<--only-first-author>|B<-f> ] ] [ B<--exclude>|B<-e> I<field> | B<--no-exclude>|B<-E> ] [ B<--set>|B<-s> I<field>B<=>I<value> ... ] [ B<--abbreviate>|B<-a> I<scheme> ... ] [ B<--pdf-file-comment>|B<-P> ] I<files>|I<directories> ...
+B<pdf-lbr-output-bib> [ B<--clipboard>|B<-c> ] [ B<--max-authors>|B<-m> I<count> [ B<--only-first-author>|B<-f> ] ] [ B<--filter>|B<-F> [I<type>B<:>]I<field>B<=>I<spec> ... ] [ B<--abbreviate>|B<-a> I<scheme> ... ] [ B<--pdf-file-comment>|B<-P> ] I<files>|I<directories> ...
 
 =head1 DESCRIPTION
 
@@ -67,17 +67,25 @@ If the number of authors is greater than I<count>, and
 
 =back
 
-=item B<--exclude>|B<-e> I<field>
+=item B<--filter>|B<-F> [I<type>B<:>]I<field>B<=>I<spec> ...
 
-Exclude the BibTeX I<field> from the output. If not set, a default list of fields given in the configuration file is excluded.
+Apply the filter I<spec> to the BibTeX I<field>. If given, I<type> applies filter only to BibTeX entries of that type. Possible I<spec> are:
 
-=item B<--no-exclude>|B<-E>
+=over 4
 
-Do not exclude any the BibTeX fields from the output.
+=item B<d>
 
-=item B<--set>|B<-s> I<field>B<=>I<value> ...
+Exclude I<field> from output.
 
-Set each BibTeX I<field> to the given I<value> before printing.
+=item B<=>I<value>
+
+Set I<field> to I<value> in output.
+
+=item B<s>B</>I<pattern>B</>I<replacement>[B</>I<pattern>B</>I<replacement>...]B</>
+
+Replace each regular expression I<pattern> with I<replacement> in output.
+
+=back
 
 =item B<--abbreviate>|B<-a> I<scheme> ...
 
@@ -112,7 +120,7 @@ PDF Librarian version @VERSION@
 =cut
 
 # handle help options
-my ($version, $help, $clipboard, $max_authors, $only_first_author, @exclude, $no_exclude, %set, @abbreviate_schemes, $pdf_file_comment);
+my ($version, $help, $clipboard, $max_authors, $only_first_author, %filter, @abbreviate_schemes, $pdf_file_comment);
 $max_authors = 0;
 $pdf_file_comment = 0;
 GetOptions(
@@ -121,16 +129,13 @@ GetOptions(
            "clipboard|c" => \$clipboard,
            "max-authors|m=i" => \$max_authors,
            "only-first-author|f" => \$only_first_author,
-           "exclude|e=s" => \@exclude,
-           "no-exclude|E" => \$no_exclude,
-           "set|s=s" => \%set,
+           "filter|F=s" => \%filter,
            "abbreviate|a=s" => \@abbreviate_schemes,
            "pdf-file-comment|P" => \$pdf_file_comment,
           ) or croak "$Script: could not parse options";
 if ($version) { print "PDF Librarian version @VERSION@\n"; exit 1; }
 pod2usage(-verbose => 2, -exitval => 1) if ($help);
 croak "$Script: --max-authors must be positive" if $max_authors < 0;
-croak "$Script: --exclude and --no-exclude are mutually exclusive" if (@exclude > 0 and $no_exclude);
 
 # get list of PDF files
 my @pdffiles = find_pdf_files(@ARGV);
@@ -139,26 +144,63 @@ croak "$Script: no PDF files to read from" unless @pdffiles > 0;
 # read BibTeX entries from PDF metadata
 my @bibentries = read_bib_from_pdf(@pdffiles);
 
-# use default exclude fields if --no-exclude is not given
-if (!$no_exclude && @exclude == 0) {
-  @exclude = @default_exclude;
-}
-if (@exclude > 0) {
-  printf STDERR "$Script: excluding BibTeX fields '%s' from output\n", join("', '", @exclude);
+# use default field filter if none given
+if (scalar(%filter) == 0) {
+  %filter = %default_filter;
 }
 
+# print field filters
+foreach my $field (sort { $filter{$a} cmp $filter{$b} || $a cmp $b } keys %filter) {
+  my $bibfield = $field;
+  my $bibselect = "BibTeX field '$bibfield'";
+  if ($bibfield =~ /^([^:]*):(.*)$/) {
+    $bibfield = $2;
+    $bibselect = "BibTeX field '$2' (in entries of type '$1')";
+  }
+  my $spec = $filter{$field};
+  if ($spec =~ /^d/) {
+    printf STDERR "$Script: excluding $bibselect from output\n";
+  } elsif ($spec =~ s/^=//) {
+    printf STDERR "$Script: setting $bibselect to value '$spec' in output\n";
+  } elsif ($spec =~ s|^s/(.*)/$|$1|) {
+    my @spec_regex = split(m|/|, $1, -1);
+    if (@spec_regex == 0 || @spec_regex % 2 != 0) {
+      croak "$Script: regular expression filter '$spec' must have form '/pattern/replacement[/pattern/replacement...]/'";
+    }
+    while (@spec_regex) {
+      my $patt = shift @spec_regex;
+      my $repl = shift @spec_regex;
+      printf STDERR "$Script: replacing regular expression '$patt' with '$repl' in output $bibselect\n";
+    }
+  } else {
+    croak "$Script: invalid spec '$bibfield=$spec' passed to --filter/-F";
+  }
+}
+
+# apply field filters
 foreach my $bibentry (@bibentries) {
-
-  # exclude BibTeX fields
-  foreach my $bibfield (@exclude) {
-    $bibentry->delete($bibfield);
+  foreach my $field (keys %filter) {
+    my $bibfield = $field;
+    if ($bibfield =~ /^([^:]*):(.*)$/) {
+      $bibfield = $2;
+      next if $bibentry->type ne $1;
+    }
+    my $spec = $filter{$field};
+    if ($spec =~ /^d/) {
+      $bibentry->delete($bibfield);
+    } elsif ($spec =~ s/^=//) {
+      $bibentry->set($bibfield, $spec);
+    } elsif ($spec =~ s|^s/(.*)/$|$1|) {
+      my @spec_regex = split(m|/|, $1, -1);
+      my $bibfieldvalue = $bibentry->get($bibfield);
+      while (@spec_regex) {
+        my $patt = shift @spec_regex;
+        my $repl = shift @spec_regex;
+        $bibfieldvalue =~ s/$patt/$repl/g;
+      }
+      $bibentry->set($bibfield, $bibfieldvalue);
+    }
   }
-
-  # set BibTeX fields
-  while (my ($bibfield, $bibvalue) = each %set) {
-    $bibentry->set($bibfield, $bibvalue);
-  }
-
 }
 
 # error if duplicate BibTeX keys are found
